+++ backend/fittoring/src/main/java/fittoring/mentoring/business/service/MentoringReservationFacadeService.java
@@ -1,12 +1,7 @@
 package fittoring.mentoring.business.service;
 
-import fittoring.mentoring.business.model.Phone;
 import fittoring.mentoring.business.model.Reservation;
-import fittoring.mentoring.business.model.Status;
 import fittoring.mentoring.business.service.dto.ReservationCreateDto;
-import fittoring.mentoring.business.service.dto.SmsReservationMessageDto;
-import fittoring.mentoring.infra.SmsMessageFormatter;
-import fittoring.mentoring.infra.SmsRestClientService;
 import fittoring.mentoring.presentation.dto.ReservationCreateResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,50 +10,17 @@ import org.springframework.stereotype.Service;
 @Service
 public class MentoringReservationFacadeService {
 
-    private static final String RESERVATION_SUBJECT = "핏토링 예약 알림";
-
     private final ReservationService reservationService;
-    private final SmsRestClientService smsRestClientService;
-    private final SmsMessageFormatter smsMessageFormatter;
+    private final ReservationNotificationService reservationNotificationService;
 
     public ReservationCreateResponse reserveMentoring(ReservationCreateDto dto) {
         Reservation reservation = reservationService.createReservation(dto);
-        String smsMessage = smsMessageFormatter.createSmsReservationMessage(SmsReservationMessageDto.of(reservation));
-        smsRestClientService.sendSms(
-                new Phone(reservation.getMentee().getPhoneNumber()),
-                smsMessage,
-                RESERVATION_SUBJECT
-        );
+        reservationNotificationService.sendReservationSmsMessage(reservation);
         return ReservationCreateResponse.from(reservation);
     }
 
     public void updateReservationStatusAndSendSms(Long reservationId, String status) {
         Reservation reservation = reservationService.updateStatus(reservationId, status);
-        sendSms(reservation, status);
+        reservationNotificationService.sendReservationStatusUpdateSmsMessage(reservation, status);
     }
-
-    private void sendSms(Reservation reservation, String updateStatus) {
-        Status status = Status.of(updateStatus);
-        String mentorName = reservation.getMentorName();
-        String context = reservation.getContext();
-        String mentorPhoneNumber = reservation.getMentorPhone();
-
-        if (status.isNotifiable()) {
-            String message = createMessage(status, mentorName, context, mentorPhoneNumber);
-            Phone menteePhone = reservation.getMentee().getPhone();
-            smsRestClientService.sendSms(menteePhone, message, RESERVATION_SUBJECT);
-        }
-    }
-
-    private String createMessage(Status updateStatus, String mentorName, String context, String mentorPhoneNumber) {
-        if (updateStatus.isApprove()) {
-            return smsMessageFormatter.approvedReservationMessage(
-                    mentorName,
-                    context,
-                    mentorPhoneNumber
-            );
-        }
-        return smsMessageFormatter.rejectedReservationMessage(mentorName);
-    }
-
 }
